# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/chat/message": {
      "post": {
        "operationId": "MessageController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send message",
        "tags": [
          "Message"
        ]
      },
      "get": {
        "operationId": "MessageController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all messages",
        "tags": [
          "Message"
        ]
      }
    },
...
    "/api/payment/stripe/webhook": {
      "post": {
        "operationId": "StripeController_handleWebhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Stripe"
        ]
      }
    },
    "/api/subscription/start-trial": {
      "post": {
        "description": "Starts a trial subscription for the authenticated user using a plan ID.",
        "operationId": "SubscriptionController_startTrial",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "planId"
                ],
                "properties": {
                  "planId": {
                    "type": "string",
                    "example": "plan_monthly_basic",
                    "description": "Internal or provider plan identifier."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trial subscription started successfully."
          },
          "400": {
            "description": "Invalid plan ID or trial not allowed."
          },
          "401": {
            "description": "Missing or invalid access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start trial subscription",
        "tags": [
          "subscription"
        ]
      }
    },
    "/api/subscription/create-product-price": {
      "post": {
        "description": "Creates a subscription product and price configuration (typically for admin/internal setup).",
        "operationId": "SubscriptionController_createProductAndPrice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductAndPriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product and price created successfully."
          },
          "400": {
            "description": "Invalid product/price payload."
          },
          "401": {
            "description": "Missing or invalid access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product and price",
        "tags": [
          "subscription"
        ]
      }
    },
    "/api/subscription/add/cards": {
      "post": {
        "description": "Adds/attaches a card to authenticated user payment profile using provider token.",
        "operationId": "SubscriptionController_addCard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card added successfully."
          },
          "400": {
            "description": "Invalid card token or product reference."
          },
          "401": {
            "description": "Missing or invalid access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add payment card",
        "tags": [
          "subscription"
        ]
      }
    },
    "/api/subscription/status": {
      "get": {
        "description": "Returns active/trial/cancelled status and relevant subscription metadata for authenticated user.",
        "operationId": "SubscriptionController_getSubscriptionStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Subscription status fetched successfully."
          },
          "401": {
            "description": "Missing or invalid access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user subscription status",
        "tags": [
          "subscription"
        ]
      }
    },
    "/api/subscription/plans": {
      "get": {
        "description": "Returns all available subscription plans for purchase/trial.",
        "operationId": "SubscriptionController_getAllPlans",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Plans fetched successfully."
          },
          "401": {
            "description": "Missing or invalid access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all plans",
        "tags": [
          "subscription"
        ]
      }
    },
    "/api/subscription/cancel": {
      "post": {
        "description": "Cancels subscription for the authenticated user. Cancellation timing depends on billing settings.",
        "operationId": "SubscriptionController_cancelSubscription",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Subscription cancelled successfully."
          },
          "400": {
            "description": "No cancellable subscription found."
          },
          "401": {
            "description": "Missing or invalid access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel current subscription",
        "tags": [
          "subscription"
        ]
      }
    }
  },
  "info": {
    "title": "Application Name api",
    "description": "Application Name api docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "receiver_id": {
            "type": "string"
          },
          "conversation_id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "receiver_id",
          "conversation_id",
          "message"
        ]
      },
      "CreateProductAndPriceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the subscription product.",
            "example": "Pro Coaching Plan"
          },
          "price": {
            "type": "number",
            "description": "Price amount in major currency unit.",
            "example": 49.99
          },
          "currency": {
            "type": "string",
            "description": "ISO currency code.",
            "example": "usd",
            "default": "usd"
          },
          "interval": {
            "type": "string",
            "description": "Billing interval for recurring subscription.",
            "enum": [
              "month",
              "year"
            ],
            "example": "month"
          },
          "interval_count": {
            "type": "number",
            "description": "Number of intervals between billings.",
            "example": 1,
            "default": 1
          },
          "product_description": {
            "type": "string",
            "description": "Optional product description shown in billing UI.",
            "example": "Access to all coaching sessions and premium content."
          },
          "price_description": {
            "type": "string",
            "description": "Optional price description shown in billing UI.",
            "example": "Billed monthly. Cancel anytime."
          },
          "trialDays": {
            "type": "number",
            "description": "Trial duration in days.",
            "example": 14
          },
          "type": {
            "type": "string",
            "description": "Internal subscription plan type.",
            "enum": [
              "FREE",
              "TRIALING",
              "BASIC",
              "PREMIUM"
            ],
            "example": "FREE"
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "AddCardDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Single-use payment token from payment provider/SDK.",
            "example": "tok_visa"
          },
          "productId": {
            "type": "string",
            "description": "Product or price identifier used to attach the card in billing flow.",
            "example": "prod_Ra9XfS8Yk2"
          }
        },
        "required": [
          "token",
          "productId"
        ]
      }
    }
  }
}